Add and rename packages
