+++ src/main/java/edu/kit/aquaplanning/grounding/BaseGrounder.java
@@ -6,6 +6,8 @@ import java.util.List;
 import java.util.Set;
 
 import edu.kit.aquaplanning.Configuration;
+import edu.kit.aquaplanning.grounding.datastructures.AtomTable;
+import edu.kit.aquaplanning.grounding.datastructures.LiftedState;
 import edu.kit.aquaplanning.model.ground.Action;
 import edu.kit.aquaplanning.model.ground.Atom;
 import edu.kit.aquaplanning.model.ground.ConditionalEffect;
