+++ src/main/java/edu/kit/aquaplanning/planning/datastructures/SearchQueue.java
@@ -1,4 +1,4 @@
-package edu.kit.aquaplanning.planning;
+package edu.kit.aquaplanning.planning.datastructures;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -10,7 +10,7 @@ import java.util.Random;
 import java.util.Set;
 import java.util.Stack;
 
-import edu.kit.aquaplanning.planning.SearchStrategy.Mode;
+import edu.kit.aquaplanning.planning.datastructures.SearchStrategy.Mode;
 import edu.kit.aquaplanning.planning.heuristic.Heuristic;
 
 /**
