+++ src/main/java/edu/kit/aquaplanning/planning/heuristic/RelaxedPathLengthHeuristic.java
@@ -2,8 +2,8 @@ package edu.kit.aquaplanning.planning.heuristic;
 
 import edu.kit.aquaplanning.model.ground.GroundPlanningProblem;
 import edu.kit.aquaplanning.model.ground.State;
-import edu.kit.aquaplanning.planning.GroundRelaxedPlanningGraph;
-import edu.kit.aquaplanning.planning.SearchNode;
+import edu.kit.aquaplanning.planning.datastructures.GroundRelaxedPlanningGraph;
+import edu.kit.aquaplanning.planning.datastructures.SearchNode;
 
 /**
  * An admissible heuristic which behaves like a simplified version of the max-cost
