+++ src/test/java/edu/kit/aquaplanning/aquaplanning/TestParallelPlanning.java
@@ -14,8 +14,8 @@ import edu.kit.aquaplanning.parsing.ProblemParser;
 import edu.kit.aquaplanning.planning.ForwardSearchPlanner;
 import edu.kit.aquaplanning.planning.Planner;
 import edu.kit.aquaplanning.planning.PortfolioParallelPlanner;
-import edu.kit.aquaplanning.planning.SimpleSatPlanner;
-import edu.kit.aquaplanning.planning.SearchStrategy.Mode;
+import edu.kit.aquaplanning.planning.datastructures.SearchStrategy.Mode;
+import edu.kit.aquaplanning.planning.sat.SimpleSatPlanner;
 import edu.kit.aquaplanning.validation.Validator;
 import junit.framework.TestCase;
